python_cert_gen/web_version/requirement.txt
blinker
cffi
click
colorama
cryptography
distlib
filelock
Flask
Flask-HTTPAuth
itsdangerous
Jinja2
MarkupSafe
platformdirs
pycparser
pyOpenSSL
Werkzeug
